complete experiment search feature
