+++ src/redux/configureStore/configureStore.js
@@ -3,7 +3,6 @@ import categoryReducer from '../movie/categories';
 import detailsReducer from '../movie/details';
 import MovieReducer from '../movie/movieAPI';
 import searchReducer from '../movie/search';
-import loneReducer from '../movie/singlemovie';
 
 const store = configureStore({
   reducer: {
@@ -11,7 +10,7 @@ const store = configureStore({
     details: detailsReducer,
     categories: categoryReducer,
     search: searchReducer,
-    lone: loneReducer
+
   },
 });
 export default store;
