+++ src/style.css
@@ -36,7 +36,7 @@ h1{
 }
 .searchBar{
     position: relative;
-    background-color: green;
+    /* background-color: green; */
     display: flex;
     flex: 1;
 }
@@ -76,8 +76,6 @@ button{
 /* ---------------home----------------  */
 .page-container{
     width: 95%;
-    background-color: darkcyan;
- 
     margin: auto;
     
 
@@ -86,10 +84,10 @@ button{
     padding-top: 28px;
     margin: auto;
     width: 80%;
-    background-color: antiquewhite;
+    /* background-color: antiquewhite; */
 }
 .category{
-    background-color: brown;
+    background-color: rgb(18, 18, 18, 0.7);
     width: 100%;
     height: 100%;
     display: inline-flex;
@@ -98,6 +96,7 @@ button{
    font-size: 39px;
    font-weight: 600;
    text-decoration: none;
+
 }
 .categories{
     padding-top: 10px;
@@ -107,13 +106,13 @@ button{
     
 }
 .col{
-    background-color: blue;
+   
     border: 1px solid white;
     width: 50%;
     height: 700px;
     display: flex;
-    /* margin-bottom: 20px; */
-    align-items: center;
+    box-shadow: 5px 10px 8px 10px #888888;
+        align-items: center;
 }
 .col span{
     border: 1px solid grey;
